refactor(lib): merge semver.ts into dependency.ts
