+++ lib/dependency.ts
@@ -1,7 +1,27 @@
 import { assertExists } from "./std/assert.ts";
 import { Mutex } from "./x/async.ts";
 import type { Maybe, Path, SemVerString } from "./types.ts";
-import { parseSemVer } from "./semver.ts";
+
+// Ref: https://semver.org/#is-there-a-suggested-regular-expression-regex-to-check-a-semver-string
+const SEMVER_REGEXP =
+  /@v?(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?/g;
+
+export function parseSemVer(
+  specifier: string,
+): Maybe<SemVerString> {
+  const match = specifier.match(SEMVER_REGEXP);
+  if (!match) {
+    return undefined;
+  }
+  if (match.length > 1) {
+    console.warn(
+      "Multiple semvers in a single specifier is not supported:",
+      specifier,
+    );
+    return undefined;
+  }
+  return match[0].slice(1) as SemVerString;
+}
 
 export interface Dependency {
   name: string;
