+++ lib/dependency_test.ts
@@ -1,6 +1,19 @@
 import { describe, it } from "./std/testing.ts";
 import { assertEquals } from "./std/assert.ts";
-import { Dependency } from "./dependency.ts";
+import { Dependency, parseSemVer } from "./dependency.ts";
+
+describe("parseSemVer", () => {
+  it("https://deno.land/std", () =>
+    assertEquals(
+      parseSemVer("https://deno.land/std@0.1.0"),
+      "0.1.0",
+    ));
+  it("https://deno.land/std (no semver)", () =>
+    assertEquals(
+      parseSemVer("https://deno.land/std"),
+      undefined,
+    ));
+});
 
 describe("parseProps()", () => {
   it("https://deno.land/std", () =>
